+++ i21Database/dbo/Tables/tblIPCommitmentPricingBalQtyStage.sql
@@ -1,10 +1,10 @@
 ﻿CREATE TABLE tblIPCommitmentPricingBalQtyStage (
 	intCommitmentPricingBalQtyStageId INT identity(1, 1)
-	,intTrxSequenceNo INT
+	,intTrxSequenceNo BIGINT
 	,strCompanyLocation NVARCHAR(6) COLLATE Latin1_General_CI_AS
 	,dtmCreatedDate DATETIME
 	,strCreatedBy NVARCHAR(50) COLLATE Latin1_General_CI_AS
-	,intLineTrxSequenceNo INT
+	,intLineTrxSequenceNo BIGINT
 	,strPricingNo NVARCHAR(50) COLLATE Latin1_General_CI_AS
 	,strERPRefNo NVARCHAR(100) COLLATE Latin1_General_CI_AS
 	,dblBalanceQty NUMERIC(18, 6)
