+++ i21Database/dbo/Tables/tblIPDemandError.sql
@@ -1,11 +1,11 @@
 ﻿CREATE TABLE dbo.tblIPDemandError (
 	intDemandErrorId int identity(1,1)
-	,intTrxSequenceNo INT
+	,intTrxSequenceNo BIGINT
 	,intActionId INT
 	,dtmCreatedDate DATETIME
 	,strCreatedBy NVARCHAR(50) COLLATE Latin1_General_CI_AS
 	,strDemandName NVARCHAR(50) COLLATE Latin1_General_CI_AS
-	,intLineTrxSequenceNo INT
+	,intLineTrxSequenceNo BIGINT
 	,strCompanyLocation NVARCHAR(50) COLLATE Latin1_General_CI_AS
 	,strItemNo NVARCHAR(50) COLLATE Latin1_General_CI_AS
 	,dblQuantity NUMERIC(18, 6)
