+++ i21Database/dbo/Tables/tblIPItemRouteDetailStage.sql
@@ -6,8 +6,8 @@
 	strManufacturingCell NVARCHAR(50) COLLATE Latin1_General_CI_AS,
 	strStorageLocation NVARCHAR(50) COLLATE Latin1_General_CI_AS,
 	strManufacturingGroup NVARCHAR(50) COLLATE Latin1_General_CI_AS,
-	intTrxSequenceNo INT,
-	intParentTrxSequenceNo INT,
+	intTrxSequenceNo BIGINT,
+	intParentTrxSequenceNo BIGINT,
 
 	CONSTRAINT [PK_tblIPItemRouteDetailStage] PRIMARY KEY (intItemRouteDetailStageId),
 	CONSTRAINT [FK_tblIPItemRouteDetailStage_tblIPItemRouteStage] FOREIGN KEY (intItemRouteStageId) REFERENCES tblIPItemRouteStage(intItemRouteStageId) ON DELETE CASCADE
