+++ i21Database/dbo/Tables/tblMFRecipeStage.sql
@@ -30,7 +30,7 @@
 	ysnImport Bit Constraint DF_tblMFRecipeStage_ysnImport Default 1,
 	ysnMailSent BIT NULL,
 	intStatusId int,
-	intTrxSequenceNo int,
+	intTrxSequenceNo BIGINT,
 	strERPRecipeNo nvarchar(50) COLLATE Latin1_General_CI_AS,
 	intActive int,
 	strSubLocationName nvarchar(50) COLLATE Latin1_General_CI_AS,
